move useObjectives to GameComponents
