+++ client/src/GameComponents/index.js
@@ -1,2 +1,3 @@
 export { GameVersion, GameVersionPicker } from './GameVersionPicker'
 export { useGameVersion, GameVersionProvider } from './useGameVersion'
+export { useObjectives } from './useObjectives'
