--- client/src/queries/index.js
@@ -1,1 +1,0 @@
-export { useObjectives } from './useObjectives'
